Optimize PixmapToBitmap a bit.

+++ intro/intro.h
@@ -20,7 +20,7 @@ static inline short TillNextKeyFrame(TrackT *track) {
 
 void FadeBlack(const PaletteT *pal, u_int start, short step);
 
-void PixmapToBitmap(BitmapT *bm, int width, int height, int depth,
+void PixmapToBitmap(BitmapT *bm, short width, short height, short depth,
                     void *pixels);
 
 #endif /* !__INTRO_H__ */
